Move BadTransactionsSource in test module
